Add typing extension
This feature simulate a click on the pinpad widget keys and commands by highlighting them when their corresponding keyboard keys are pressed. For example, if the numpad key 1 is pressed, or more precisely if the key is down, the pinpad key 1 assume the active state, and come back to its normal state when the key is up.