Add the `value` option
The `value` option is used to store the last confirmed value. If the user edits and confirms the value, the pinpad input value is stored in the `value` option. If the user edits and presses the cancel button or leave the pinpad, the pinpad input value is reset to the `value` option.